Add Keyboard and WIP Menu Modal
Added the functioning virtual keyboard for use in mobile and for
mostly just keeping track of used letters in a PC browser.
Keyboard correctly interacts with a single instance of a Word
component, appending and deleting as fit but no 'ENTER'
functionality yet.

Menu is implemented using react-modal with letters and attempts
able to be configured using dropdowns of values between 3 and 10.