Add Feedback and End Screen
Added parsing of user input to check for letter matches, including
dynamic colouring of keyboard keys and word tiles to show results.

The ENTER button will now also test for completion of the game, be
it from guessing the word correctly or running out of guesses.
When this happens, a game over UI will come up, showing results
as well as buttons to change the settings or play again.

The only thing left to do is to enable physical keyboard entry.